+++ src/main/java/co/edu/utp/misiontic2022/c2/cdiaz/model/vo/DeudasPorProyectoVo.java
@@ -3,7 +3,7 @@ package co.edu.utp.misiontic2022.c2.cdiaz.model.vo;
 public class DeudasPorProyectoVo {
 
     private Integer id_Proyecto;
-    private Integer valor;
+    private Double valor;
     
     public Integer getid_Proyecto() {
         return  id_Proyecto;
@@ -13,17 +13,16 @@ public class DeudasPorProyectoVo {
         this.id_Proyecto = id_Proyecto;
     }
     
-    public Integer getValor() {
+    public Double getValor() {
         return valor;
     }
     
-    public void setValor(Integer valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
 
     @Override
     public String toString() {
         return "DeudasPorProyectoVo [id_Proyecto=" + id_Proyecto + ", valor=" + valor + "]";
-    }
-       
+    }       
 }
